+++ Headers/cpu_ops.hpp
@@ -4,6 +4,6 @@
 #include "globals.hpp"
 
 namespace cpu {
-    std::vector<double> matmul(const std::vector<double> &A, const std::vector<double> &B, int K, int X, int Y);
-    void softmax(const std::vector<double>& input, double* output, int N, int C);
+    std::vector<float> matmul(const std::vector<float> &A, const std::vector<float> &B, int K, int X, int Y);
+    void softmax(const std::vector<float>& input, float* output, int N, int C);
 }
