+++ Headers/cuda/cuda_launchers.h
@@ -3,8 +3,8 @@
 #include "globals.hpp"
 // TODO: rename file to cuda_kernels?
 
-void launch_matmul(const double* d_A, const double* d_B, double* d_C, int K, int X, int Y);
-void launch_binary_op(const char op, const double* input_A, const double* input_B, double* output, int size);
-void launch_softmax2(const double* input_A, double* output, int N, int C);
-int launch_reduction(const ReductionOp op, const double* input, double* output, int size);
-void launch_full_reduction(const ReductionOp op, const double* input, double* output, int size);
+void launch_matmul(const float* d_A, const float* d_B, float* d_C, int K, int X, int Y);
+void launch_binary_op(const char op, const float* input_A, const float* input_B, float* output, int size);
+void launch_softmax2(const float* input_A, float* output, int N, int C);
+int launch_reduction(const ReductionOp op, const float* input, float* output, int size);
+void launch_full_reduction(const ReductionOp op, const float* input, float* output, int size);
