+++ Source/loss.cpp
@@ -11,7 +11,7 @@ Tensor_ptr MSELoss(Tensor_ptr input, Tensor_ptr target)
     result = result * result;
     result = result->sum();
 
-    Tensor_ptr div = Tensor::init({1}, std::vector<double>{static_cast<double>(input->get_shape(0))}, input->get_device());
+    Tensor_ptr div = Tensor::init({1}, std::vector<float>{static_cast<float>(input->get_shape(0))}, input->get_device());
     result = result / div;
     return result;
 }
